fixed prior pr comments
