+++ src/main/java/com/google/codeu/servlets/AddPlaceServlet.java
@@ -1,6 +1,5 @@
 package com.google.codeu.servlets;
 
-import com.google.api.Http;
 import javax.servlet.ServletException;
 import javax.servlet.annotation.WebServlet;
 import javax.servlet.http.HttpServlet;
@@ -8,30 +7,17 @@ import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import java.io.IOException;
 import java.util.Properties;
-import com.google.appengine.api.users.UserService;
-import com.google.appengine.api.users.UserServiceFactory;
-import org.jsoup.Jsoup;
-import org.jsoup.safety.Whitelist;
 
-@WebServlet("/api/place")
-public class AddPlaceServlet extends HttpServlet {
+@WebServlet("/place/add")
+public class AddPlaceServlet extends HttpServlet{
     @Override
-    public void doPost(HttpServletRequest request, HttpServletResponse response)
-        throws IOException {
-        UserService userService = UserServiceFactory.getUserService();
-        if (!userService.isUserLoggedIn()) {
-            response.sendRedirect("/");
-            return;
-        }
+    public void doGet(HttpServletRequest request, HttpServletResponse response)
+            throws IOException, ServletException {
+        Properties props = ServletUtils.GetProperties();
+        request.setAttribute("TITLE", "Add Place");
+        request.setAttribute(
+                "HEAD_HTML", "");
 
-        String user = userService.getCurrentUser().getEmail();
-        // Get the message entered by the user.
-        String userXCoordText =
-                Jsoup.clean(request.getParameter("x_coord"), Whitelist.none());
-
-
-        // Get the URL of the image that the user uploaded to Blobstore.
-        String imageUrl = BlobstoreServlet.getUploadedFileUrl(request, "image");
+        ServletUtils.RenderReact("addPlace", request, response);
     }
 }
-
